Implementação de seguranças nas apis

+++ backend/handlers/rest/deleteBoard.js
@@ -1,15 +1,23 @@
 const { getBoardDb, deleteBoardDb } = require('../../services/database/dynamoService');
 const config = require('../../config');
 const log = require('../../utils/logger');
+const { getAuthenticatedUserId } = require('../../utils/auth');
 
 exports.handler = async (event) => {
   log.setCorrelationId(event.requestContext.requestId);
   const { boardId } = event.pathParameters;
+  const userId = getAuthenticatedUserId(event);
 
-  log.debug('Delete board request', { boardId });
+  log.debug('Delete board request', { boardId, userId });
 
   try {
     const board = await getBoardDb(config.TABLE_BOARD, boardId);
+
+    if (board.creatorId !== userId) {
+      log.warn('Tentativa de excluir board de outro usuário', { boardId, userId, creatorId: board.creatorId });
+      return { statusCode: 403, body: JSON.stringify({ error: 'Você não tem permissão para excluir este board' }) };
+    }
+
     await deleteBoardDb(config.TABLE_BOARD, boardId, board.createdAt);
     log.info('Board deleted', { boardId, boardName: board.boardName, creatorId: board.creatorId });
     return { statusCode: 204, body: '' };
